feat: creating home_ui module
