+++ app/src/main/java/br/estudo/everty/app_weather/app/WeatherApp.kt
@@ -2,7 +2,6 @@ package br.estudo.everty.app_weather.app
 
 import android.app.Application
 import br.estudo.everty.app_weather.app.di.HomeModuleInit
-import br.estudo.everty.app_weather.home.di.injectHomeModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
 
@@ -13,8 +12,6 @@ class WeatherApp: Application() {
         startKoin {
             androidContext(this@WeatherApp)
         }
-
-        injectHomeModule()
         initFeatures()
     }
