--- app/src/main/java/br/estudo/everty/app_weather/home/ui/HomeActivity.kt
@@ -1,7 +1,0 @@
-package br.estudo.everty.app_weather.home.ui
-
-import br.estudo.everty.app_weather.R
-import br.com.everty.shared.utils.activity.BaseActivity
-
-
-class HomeActivity: BaseActivity(R.layout.activity_home)
