--- app/src/main/java/br/estudo/everty/app_weather/home/ui/screens/events/HomeEvents.kt
@@ -1,6 +1,0 @@
-package br.estudo.everty.app_weather.home.ui.screens.events
-
-interface HomeEvents {
-    fun onClickError()
-    fun onFilterSelected(isDaily: Boolean)
-}
