+++ app/src/main/java/br/estudo/everty/app_weather/splash/SplashActivity.kt
@@ -3,8 +3,8 @@ package br.estudo.everty.app_weather.splash
 import android.content.Intent
 import android.os.Bundle
 import androidx.activity.compose.setContent
+import br.com.everty.home.home_ui.HomeActivity
 import br.com.everty.shared.presentation.design_system.theme.AppWeatherTheme
-import br.estudo.everty.app_weather.home.ui.HomeActivity
 import br.com.everty.shared.utils.activity.BaseActivity
