+++ features/home/home_ui/src/main/AndroidManifest.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android">
+    <application>
+        <activity
+            android:name="br.com.everty.home.home_ui.HomeActivity"
+            android:exported="true"
+            android:theme="@style/Theme.AppWeather">
+        </activity>
+    </application>
+
+</manifest>
