+++ features/home/home_ui/src/main/java/br/com/everty/home/home_ui/HomeActivity.kt
@@ -1,0 +1,6 @@
+package br.com.everty.home.home_ui
+
+import br.com.everty.shared.utils.activity.BaseActivity
+
+
+class HomeActivity: BaseActivity(R.layout.activity_home)
