+++ features/home/home_ui/src/main/java/br/com/everty/home/home_ui/events/HomeEvents.kt
@@ -1,0 +1,6 @@
+package br.com.everty.home.home_ui.events
+
+interface HomeEvents {
+    fun onClickError()
+    fun onFilterSelected(isDaily: Boolean)
+}
