+++ settings.gradle.kts
@@ -20,3 +20,4 @@ include(":shared:network")
 include(":shared:utils")
 include(":features:home:home_data")
 include(":features:home:home_domain")
+include(":features:home:home_ui")
